Add delete button confirm

+++ app/src/main/java/com/example/pricetag/templates/index/IndexAdapter.java
@@ -1,6 +1,8 @@
 package com.example.pricetag.templates.index;
 
+import android.app.AlertDialog;
 import android.content.Context;
+import android.content.DialogInterface;
 import android.os.Bundle;
 import android.view.LayoutInflater;
 import android.view.View;
@@ -53,15 +55,25 @@ public class IndexAdapter extends RecyclerView.Adapter<IndexAdapter.ViewHolder> 
         int id = items.get(position).getId();
         holder.setId(id);
 
-        holder.editActionButton.setOnClickListener(new View.OnClickListener() {
-            public void onClick(View v) {
-                holder.handleEditClickEvent(v);
-            }
-        });
+        holder.editActionButton.setOnClickListener(holder::handleEditClickEvent);
 
         holder.deleteActionButton.setOnClickListener(v -> {
-            int itemId = holder.getId();
-            fragment.deleteData(itemId);
+
+            new AlertDialog.Builder(context)
+                    .setTitle("Delete entry")
+                    .setMessage("Are you sure you want to delete this entry?")
+
+                    // Specifying a listener allows you to take an action before dismissing the dialog.
+                    // The dialog is automatically dismissed when a dialog button is clicked.
+                    .setPositiveButton(android.R.string.yes, (dialog, which) -> {
+                        int itemId = holder.getId();
+                        fragment.deleteData(itemId);
+                    })
+
+                    // A null listener allows the button to dismiss the dialog and take no further action.
+                    .setNegativeButton(android.R.string.no, null)
+                    .setIcon(android.R.drawable.ic_dialog_alert)
+                    .show();
 
         });
